minio, kernel-plancksater config: set kp's minio config to point to public minio api url
this requires KP_OBJECT_STORE_SECURE to be False and cert check to be disabled so that the minio client in kp can make calls to the public minio endpoint.

This is required to generate signed url signatures with public hostname encypted in the signature. Else we get a "SignatureDoesNotMatch" error.  i.e If the signed url is generated from cluster internal DNS and a URL rewrite is performed to the public minio address, the signature is considered bad and signed-url based uploads/downloads will fail.